requirements.txt
PyYAML==5.3
redis==3.3.11
